+++ packages/config-webpack4/src/index.ts
@@ -1,0 +1,12 @@
+import {
+  DevelopmentConfiguration as webpack4_DevelopmentConfiguration,
+  ProductionConfiguration as webpack4_ProductionConfiguration,
+  getWebpackConfiguration as getWebpack4Configuration,
+} from './configurations';
+
+export type { webpack4_DevelopmentConfiguration, webpack4_ProductionConfiguration };
+
+export {
+  // webpack
+  getWebpack4Configuration,
+};