+++ packages/toolkit/CHANGELOG.md
@@ -3,6 +3,23 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.2.0](https://github.com/wearehumblebee/honeycomb/compare/@humblebee/toolkit@0.1.3...@humblebee/toolkit@0.2.0) (2021-04-06)
+
+
+### Features
+
+* **build-configs:** break down toolkit configuration into specific packages ([5e232ff](https://github.com/wearehumblebee/honeycomb/commit/5e232ffbbb6870d6ac8ebbda9cd5659692aa5065))
+
+
+### BREAKING CHANGES
+
+* **build-configs:** @humblebee/toolkit no longer provides build configurations for babel and webpack.
+(Those can be imported respectively from @humblebee/config-babel7 and @humblebee/config-webpack4)
+
+
+
+
+
 ## [0.1.3](https://github.com/wearehumblebee/honeycomb/compare/@humblebee/toolkit@0.1.2...@humblebee/toolkit@0.1.3) (2021-03-16)
 
 **Note:** Version bump only for package @humblebee/toolkit